pyproject.toml
[project]
name = "simple-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
